requirements-dev.txt
# Stubs for Blender API
fake-bpy-module
